+++ day_13/java/app/src/main/java/tkrippes/com/github/adventofcode2024/day13/Position.java
@@ -1,6 +1,6 @@
 package tkrippes.com.github.adventofcode2024.day13;
 
-public record Position(int x, int y) {
+public record Position(long x, long y) {
     @Override
     public String toString() {
         return "X=" + x + ", Y=" + y;
